Antlr C++ parser/lexer code is generated from SQL.g4 if it is missing.
Quite a bit of trickery is needed for this to happen:
The variable ANTLR_GEN uses a ':=' for assignment, instead of a '='.
The variable is resolved only once, at the start of the Make run,
before any targets are run (i.e. before any compilation).
The variable uses a subshell to check for an empty 'antlr_output' directory
and runs the Java Antlr .jar file to generate a C++ parser/lexer from
the grammar in SQL.g4.
If the 'antlr_output' directory is not empty, then java is not run.

Note: This Makefile does not yet resolve dependencies on SQL.g4.
      If SQL.g4 is modified, then a 'make cleanall' and 'make' is
      required to generate a new parser/lexer from SQL.g4.

VPATH can be separated with spaces instead of colons. More portable.

Use BIN_DIR variable where possible.